+++ CHANGELOG.md
@@ -5,7 +5,20 @@ All notable changes to `exonet-api-python` will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
-[Compare 2.0.0 - Unreleased](https://github.com/exonet/exonet-api-python/compare/2.0.0...master)
+[Compare 3.0.0 - Unreleased](https://github.com/exonet/exonet-api-python/compare/3.0.0...master)
+
+## [3.0.0](https://github.com/exonet/exonet-api-python/releases/tag/3.0.0) - 2020-09-11
+[Compare 2.1.0 - 3.0.0](https://github.com/exonet/exonet-api-python/compare/2.1.0...3.0.0)
+### Breaking
+- When multiple resources are returned from the API, an instance of `ApiResourceSet` is returned instead of a list. This class is traversable so unless the code does specific `list` things or type checks, no changes are necessary.
+
+### Added
+- Add the `total()` method to resource sets to get the total number of resources (and not only the number of resources in the current resource set).
+- Add `next_page`, `previous_page`, `first_page` and `last_page` methods to the `ApiResourceSet` for easy loading of paginated resource sets.
+- Add a `get_recursive` method to the `RequestBuilder` to get the resource set including recursively the resource sets from the following pages.
+
+### Removed
+- The `store` method for creating `POST` requests. (Deprecated since 2.0.0)
 
 ## [2.1.0](https://github.com/exonet/exonet-api-python/releases/tag/2.1.0) - 2019-11-19
 [Compare 2.0.0 - 2.1.0](https://github.com/exonet/exonet-api-python/compare/2.0.0...2.1.0)
@@ -18,7 +31,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - The `Api` prefix has been added from the following classes for consistency:
   - `Resource` --> `ApiResource`
   - `ResourceIdentifier` --> `ApiResourceIdentifier`
-  
+
 ### Added
 - Support for `PATCH` and `DELETE` requests.